requirements.txt
flask>=1.1.2
flask-cors>=3.0.8
flask_restful>=0.3.8
google-api-python-client>=1.10.0
google-cloud-translate>=2.0.1
grpcio>=1.30.0
numpy>=1.19.1
oauth2-client>=1.2.1
Pillow==6.2.2
pytest>=5.4.3
pytest-flask>=1.0.0
python-dotenv>=0.14.0
requests>=2.24.0
spacy>=2.3.2
stimson-web-scraper>=0.0.58
waitress>=1.4.4
wheel>=0.34.2
